Migrating from llm-wrapper to allms. All code refactored, docs and README.md updated
