+++ allms/constants/azure.py
