+++ allms/defaults/azure_defaults.py
