+++ allms/domain/configuration.py
