+++ allms/domain/enumerables.py
