+++ allms/domain/input_data.py
