+++ allms/domain/prompt_dto.py
