+++ docs/usage/basic.md
@@ -2,12 +2,12 @@
 
 ## Single Query 
 
-In the simplest approach you just need to pass a prompt and the model will provide a response for it.  
+In the simplest approach you just need to pass a prompt and the model will provide a response for it.
 
 ```python
-from llm_wrapper.models import AzureOpenAIModel
-from llm_wrapper.domain.configuration import AzureOpenAIConfiguration
-from llm_wrapper.domain.response import ResponseData
+from allms.models import AzureOpenAIModel
+from allms.domain.configuration import AzureOpenAIConfiguration
+from allms.domain.response import ResponseData
 
 configuration = AzureOpenAIConfiguration(
     api_key="<OPENAI_API_KEY>",
@@ -22,7 +22,7 @@ model = AzureOpenAIModel(config=configuration)
 response = model.generate("What is the capital of Poland?")
 print(response)
 
-#[ResponseData(response='The capital of Poland is Warsaw.', input_data=None, number_of_prompt_tokens=7, number_of_generated_tokens=7, error=None)]
+# [ResponseData(response='The capital of Poland is Warsaw.', input_data=None, number_of_prompt_tokens=7, number_of_generated_tokens=7, error=None)]
 ```
 
 As a response you'll get `List[ResponseData]`, where the first element will contain response from the model in the
